business-card-scanner/scripts/requirements.txt
pytesseract>=0.3.10
pillow>=10.0.0
opencv-python>=4.8.0
pandas>=2.0.0
numpy>=1.24.0
